Completed the base compilation of input participant data + added expandable validation
Completed the system which would compile the participant data inputed by the player and send it to the game starter. Also added expandable validation to check if some starting conditions have been me.

PS
the compilation process happens from the top (teams) towards the bottom (players) so every player gets check on attachment to a spawnpoint. The ones that pass get added to the compiled array of base participants.