Finished Player type-changing limitations + Added player and team name changing + added team compiling
Completed the limitation system for player types (you can now specify what is the maximum number of sertain types of players). Added player and team names that can be changed vie input fields and are saved to their respective objects and later compiled.  Teams now compile into baseteam arrays (unfortunately all of them do: teams that are not used are not excluded yet)