Added the keypoint system to assist in AI targeting
The tower now has a number of keypoints that represent structuraly important points in the tower and can recalculate their importance. This way when the kypoint is hit and the structure related to it is destroyed it will be marked as destroyed and the AI  will know to ignore it. Also added the tower class that will sit on every tower and help with tower to tower interaction.