Finished the new tower controll room and added a gauge display.
Finished the new tower controll room design (that is lookated in the tower but you can still climb on top of it). Added a gauge display that helps with the intuitivness of aiming. Added sound to the armored roo. Added trapdoor wall models and one prefab. Set up the climbing out of the control room. Added a fence. Fixed a bug where the cannon room joints would break.