Merged the player and AI towers into a single prefab. Reworked the tower component ot support toggling the ai and player movement from tower to tower.
Both the player and the AI tower are now a single prefab that has a console and an intenral AI that is toggled on by the tower when there are no players. The screen also stops rendering when there are no players. The tower is simply a structure managment system while the playable tower manages the players in a tower and the AI toggling it on and off.  The playable tower component now has a public list of players that, when changed, updates the tower. with new players.