Added tower color indicators
Added tower color indicators that appear as giant rectangles over every tower that are only visible to the radar. Their color is set in the gamestarter which call the set color function on the tower. When the tower is destroyed it resets their color to the one specified in its component.