Added a couple of new maps, reworked the organization
Added canyon and a couple other maps, reworked organization. Now everything related to the map visuals is stored in the volumes folder under every map.

To create a new map duplicate one of those folders and add the visual prefab into the map, replacing the volumes in the visual prefab with the ones just duplicated and maybe adding new ones
There is also a general volume folder for default visual prefab settings like exposure that do not need to be replaced in every map.