Finished the communication system, completed the rooms-tower structure, added all the references. Started work on the spawn system
The communication system is now finished. It consists of a playable tower and 4 rooms that communicate between each other passing regular info via the reference system and specific info like the cannon to the reloader via methods. Added the spawn system (but haven't tested it yet). The only thing to do now is to make a tower creation screen and make some new rooms.